third_party/blink/web_tests/platform/mac/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/elementwise_unary.https.any.worker_gpu-expected.txt

This is a testharness.js-based test.
Found 64 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] abs float32 1D constant tensor
  assert_true: assert_array_approx_equals_ulp: test abs float32 actual 49.8125 should be close enough to expected 49.837242126464844 by the acceptable 0 ULP distance, but they have 6486 ULP distance expected true got false
[FAIL] abs float32 1D tensor
  assert_true: assert_array_approx_equals_ulp: test abs float32 actual 49.8125 should be close enough to expected 49.837242126464844 by the acceptable 0 ULP distance, but they have 6486 ULP distance expected true got false
[FAIL] abs float32 2D tensor
  assert_true: assert_array_approx_equals_ulp: test abs float32 actual 49.8125 should be close enough to expected 49.837242126464844 by the acceptable 0 ULP distance, but they have 6486 ULP distance expected true got false
[FAIL] abs float32 3D tensor
  assert_true: assert_array_approx_equals_ulp: test abs float32 actual 49.8125 should be close enough to expected 49.837242126464844 by the acceptable 0 ULP distance, but they have 6486 ULP distance expected true got false
[FAIL] abs float32 4D tensor
  assert_true: assert_array_approx_equals_ulp: test abs float32 actual 49.8125 should be close enough to expected 49.837242126464844 by the acceptable 0 ULP distance, but they have 6486 ULP distance expected true got false
[FAIL] abs float32 5D tensor
  assert_true: assert_array_approx_equals_ulp: test abs float32 actual 49.8125 should be close enough to expected 49.837242126464844 by the acceptable 0 ULP distance, but they have 6486 ULP distance expected true got false
[FAIL] ceil float32 1D constant tensor
  assert_true: assert_array_approx_equals_ulp: test ceil float32 actual 5 should be close enough to expected 6 by the acceptable 0 ULP distance, but they have 2097152 ULP distance expected true got false
[FAIL] ceil float32 1D tensor
  assert_true: assert_array_approx_equals_ulp: test ceil float32 actual 5 should be close enough to expected 6 by the acceptable 0 ULP distance, but they have 2097152 ULP distance expected true got false
[FAIL] ceil float32 2D tensor
  assert_true: assert_array_approx_equals_ulp: test ceil float32 actual 5 should be close enough to expected 6 by the acceptable 0 ULP distance, but they have 2097152 ULP distance expected true got false
[FAIL] ceil float32 3D tensor
  assert_true: assert_array_approx_equals_ulp: test ceil float32 actual 5 should be close enough to expected 6 by the acceptable 0 ULP distance, but they have 2097152 ULP distance expected true got false
[FAIL] ceil float32 4D tensor
  assert_true: assert_array_approx_equals_ulp: test ceil float32 actual 5 should be close enough to expected 6 by the acceptable 0 ULP distance, but they have 2097152 ULP distance expected true got false
[FAIL] ceil float32 5D tensor
  assert_true: assert_array_approx_equals_ulp: test ceil float32 actual 5 should be close enough to expected 6 by the acceptable 0 ULP distance, but they have 2097152 ULP distance expected true got false
[FAIL] cos float32 1D constant tensor
  assert_array_approx_equals: test cos float32 property 1, expected 0.43789437413215637 +/- 0.0009765625, expected 0.43789437413215637 but got 0.443359375
[FAIL] cos float32 1D tensor
  assert_array_approx_equals: test cos float32 property 1, expected 0.43789437413215637 +/- 0.0009765625, expected 0.43789437413215637 but got 0.443359375
[FAIL] cos float32 2D tensor
  assert_array_approx_equals: test cos float32 property 1, expected 0.43789437413215637 +/- 0.0009765625, expected 0.43789437413215637 but got 0.443359375
[FAIL] cos float32 3D tensor
  assert_array_approx_equals: test cos float32 property 1, expected 0.43789437413215637 +/- 0.0009765625, expected 0.43789437413215637 but got 0.443359375
[FAIL] cos float32 4D tensor
  assert_array_approx_equals: test cos float32 property 1, expected 0.43789437413215637 +/- 0.0009765625, expected 0.43789437413215637 but got 0.443359375
[FAIL] cos float32 5D tensor
  assert_array_approx_equals: test cos float32 property 1, expected 0.43789437413215637 +/- 0.0009765625, expected 0.43789437413215637 but got 0.443359375
[FAIL] exp float32 1D constant tensor
  assert_true: assert_array_approx_equals_ulp: test exp float32 actual 1.4072265625 should be close enough to expected 1.4079625606536865 by the acceptable 32 ULP distance, but they have 6174 ULP distance expected true got false
[FAIL] exp float32 1D tensor
  assert_true: assert_array_approx_equals_ulp: test exp float32 actual 1.4072265625 should be close enough to expected 1.4079625606536865 by the acceptable 32 ULP distance, but they have 6174 ULP distance expected true got false
[FAIL] exp float32 2D tensor
  assert_true: assert_array_approx_equals_ulp: test exp float32 actual 1.4072265625 should be close enough to expected 1.4079625606536865 by the acceptable 32 ULP distance, but they have 6174 ULP distance expected true got false
[FAIL] exp float32 3D tensor
  assert_true: assert_array_approx_equals_ulp: test exp float32 actual 1.4072265625 should be close enough to expected 1.4079625606536865 by the acceptable 32 ULP distance, but they have 6174 ULP distance expected true got false
[FAIL] exp float32 4D tensor
  assert_true: assert_array_approx_equals_ulp: test exp float32 actual 1.4072265625 should be close enough to expected 1.4079625606536865 by the acceptable 32 ULP distance, but they have 6174 ULP distance expected true got false
[FAIL] exp float32 5D tensor
  assert_true: assert_array_approx_equals_ulp: test exp float32 actual 1.4072265625 should be close enough to expected 1.4079625606536865 by the acceptable 32 ULP distance, but they have 6174 ULP distance expected true got false
[FAIL] identity float32 0D scalar
  assert_true: assert_array_approx_equals_ulp: test identity float32 actual -4.2734375 should be close enough to expected -4.273642539978027 by the acceptable 0 ULP distance, but they have 430 ULP distance expected true got false
[FAIL] identity float32 1D constant tensor
  assert_true: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by the acceptable 0 ULP distance, but they have 3691 ULP distance expected true got false
[FAIL] identity float32 1D tensor
  assert_true: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by the acceptable 0 ULP distance, but they have 3691 ULP distance expected true got false
[FAIL] identity float32 2D tensor
  assert_true: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by the acceptable 0 ULP distance, but they have 3691 ULP distance expected true got false
[FAIL] identity float32 3D tensor
  assert_true: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by the acceptable 0 ULP distance, but they have 3691 ULP distance expected true got false
[FAIL] identity float32 4D tensor
  assert_true: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by the acceptable 0 ULP distance, but they have 3691 ULP distance expected true got false
[FAIL] identity float32 5D tensor
  assert_true: assert_array_approx_equals_ulp: test identity float32 actual 0.37744140625 should be close enough to expected 0.377551406621933 by the acceptable 0 ULP distance, but they have 3691 ULP distance expected true got false
[FAIL] log float32 positive 1D constant tensor
  assert_array_approx_equals: test log float32 property 0, expected 4.15615177154541 +/- 0.0009765625, expected 4.15615177154541 but got 4.15234375
[FAIL] log float32 positive 1D tensor
  assert_array_approx_equals: test log float32 property 0, expected 4.15615177154541 +/- 0.0009765625, expected 4.15615177154541 but got 4.15234375
[FAIL] log float32 positive 2D tensor
  assert_array_approx_equals: test log float32 property 0, expected 4.15615177154541 +/- 0.0009765625, expected 4.15615177154541 but got 4.15234375
[FAIL] log float32 positive 3D tensor
  assert_array_approx_equals: test log float32 property 0, expected 4.15615177154541 +/- 0.0009765625, expected 4.15615177154541 but got 4.15234375
[FAIL] log float32 positive 4D tensor
  assert_array_approx_equals: test log float32 property 0, expected 4.15615177154541 +/- 0.0009765625, expected 4.15615177154541 but got 4.15234375
[FAIL] log float32 positive 5D tensor
  assert_array_approx_equals: test log float32 property 0, expected 4.15615177154541 +/- 0.0009765625, expected 4.15615177154541 but got 4.15234375
[FAIL] neg float32 1D constant tensor
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': This operator (neg) is not implemented."
[FAIL] neg float32 1D tensor
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': This operator (neg) is not implemented."
[FAIL] neg float32 2D tensor
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': This operator (neg) is not implemented."
[FAIL] neg float32 3D tensor
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': This operator (neg) is not implemented."
[FAIL] neg float32 4D tensor
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': This operator (neg) is not implemented."
[FAIL] neg float32 5D tensor
  promise_test: Unhandled rejection with value: object "NotSupportedError: Failed to execute 'build' on 'MLGraphBuilder': This operator (neg) is not implemented."
[FAIL] reciprocal float32 0D scalar
  assert_true: assert_array_approx_equals_ulp: test reciprocal float32 actual -0.1068115234375 should be close enough to expected -0.10679777711629868 by the acceptable 2 ULP distance, but they have 1845 ULP distance expected true got false
[FAIL] reciprocal float32 1D constant tensor
  assert_true: assert_array_approx_equals_ulp: test reciprocal float32 actual -0.9892578125 should be close enough to expected -0.9890353083610535 by the acceptable 2 ULP distance, but they have 3733 ULP distance expected true got false
[FAIL] reciprocal float32 1D tensor
  assert_true: assert_array_approx_equals_ulp: test reciprocal float32 actual -0.9892578125 should be close enough to expected -0.9890353083610535 by the acceptable 2 ULP distance, but they have 3733 ULP distance expected true got false
[FAIL] reciprocal float32 2D tensor
  assert_true: assert_array_approx_equals_ulp: test reciprocal float32 actual -0.9892578125 should be close enough to expected -0.9890353083610535 by the acceptable 2 ULP distance, but they have 3733 ULP distance expected true got false
[FAIL] reciprocal float32 3D tensor
  assert_true: assert_array_approx_equals_ulp: test reciprocal float32 actual -0.9892578125 should be close enough to expected -0.9890353083610535 by the acceptable 2 ULP distance, but they have 3733 ULP distance expected true got false
[FAIL] reciprocal float32 4D tensor
  assert_true: assert_array_approx_equals_ulp: test reciprocal float32 actual -0.9892578125 should be close enough to expected -0.9890353083610535 by the acceptable 2 ULP distance, but they have 3733 ULP distance expected true got false
[FAIL] reciprocal float32 5D tensor
  assert_true: assert_array_approx_equals_ulp: test reciprocal float32 actual -0.9892578125 should be close enough to expected -0.9890353083610535 by the acceptable 2 ULP distance, but they have 3733 ULP distance expected true got false
[FAIL] sin float32 1D constant tensor
  assert_array_approx_equals: test sin float32 property 0, expected -0.946033775806427 +/- 0.0009765625, expected -0.946033775806427 but got -0.935546875
[FAIL] sin float32 1D tensor
  assert_array_approx_equals: test sin float32 property 0, expected -0.946033775806427 +/- 0.0009765625, expected -0.946033775806427 but got -0.935546875
[FAIL] sin float32 2D tensor
  assert_array_approx_equals: test sin float32 property 0, expected -0.946033775806427 +/- 0.0009765625, expected -0.946033775806427 but got -0.935546875
[FAIL] sin float32 3D tensor
  assert_array_approx_equals: test sin float32 property 0, expected -0.946033775806427 +/- 0.0009765625, expected -0.946033775806427 but got -0.935546875
[FAIL] sin float32 4D tensor
  assert_array_approx_equals: test sin float32 property 0, expected -0.946033775806427 +/- 0.0009765625, expected -0.946033775806427 but got -0.935546875
[FAIL] sin float32 5D tensor
  assert_array_approx_equals: test sin float32 property 0, expected -0.946033775806427 +/- 0.0009765625, expected -0.946033775806427 but got -0.935546875
[FAIL] sqrt float32 0D scalar
  assert_true: assert_array_approx_equals_ulp: test sqrt float32 actual 2 should be close enough to expected 2.001115560531616 by the acceptable 1 ULP distance, but they have 4679 ULP distance expected true got false
[FAIL] sqrt float32 1D constant tensor
  assert_true: assert_array_approx_equals_ulp: test sqrt float32 actual 2.69140625 should be close enough to expected 2.693697690963745 by the acceptable 1 ULP distance, but they have 9611 ULP distance expected true got false
[FAIL] sqrt float32 1D tensor
  assert_true: assert_array_approx_equals_ulp: test sqrt float32 actual 2.69140625 should be close enough to expected 2.693697690963745 by the acceptable 1 ULP distance, but they have 9611 ULP distance expected true got false
[FAIL] sqrt float32 2D tensor
  assert_true: assert_array_approx_equals_ulp: test sqrt float32 actual 2.69140625 should be close enough to expected 2.693697690963745 by the acceptable 1 ULP distance, but they have 9611 ULP distance expected true got false
[FAIL] sqrt float32 3D tensor
  assert_true: assert_array_approx_equals_ulp: test sqrt float32 actual 2.69140625 should be close enough to expected 2.693697690963745 by the acceptable 1 ULP distance, but they have 9611 ULP distance expected true got false
[FAIL] sqrt float32 4D tensor
  assert_true: assert_array_approx_equals_ulp: test sqrt float32 actual 2.69140625 should be close enough to expected 2.693697690963745 by the acceptable 1 ULP distance, but they have 9611 ULP distance expected true got false
[FAIL] sqrt float32 5D tensor
  assert_true: assert_array_approx_equals_ulp: test sqrt float32 actual 2.69140625 should be close enough to expected 2.693697690963745 by the acceptable 1 ULP distance, but they have 9611 ULP distance expected true got false
[FAIL] tan float32 1D constant tensor
  assert_array_approx_equals: test tan float32 property 0, expected -0.8582430481910706 +/- 0.0009765625, expected -0.8582430481910706 but got -0.8763164281845093
Harness: the test ran to completion.
